requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
cycler==0.11.0
dacite==1.6.0
Deprecated==1.2.13
discord-pretty-help==1.3.3
discord.py==1.7.3
dnspython==2.1.0
fonttools==4.28.5
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
multidict==5.2.0
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
pokemontcgsdk==3.4.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
pymongo==4.0.1
PyNaCl==1.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
wrapt==1.13.3
yarl==1.7.2
